requirements.txt
python-binance
websockets~=10.4
prettytable~=3.5.0
pandas